Ported the mod to 1.20.1

+++ src/main/java/com/solegendary/reignofnether/building/buildings/monsters/SlimePit.java
@@ -91,6 +91,6 @@ public class SlimePit extends ProductionBuilding {
     }
 
     public BlockPos getDefaultOutdoorSpawnPoint() {
-        return getMinCorner(this.blocks).offset(-spawnRadiusOffset + 4, 0, -spawnRadiusOffset + 9);
+        return getMinCorner(this.blocks).offset((int) (-spawnRadiusOffset + 4), 0, (int) (-spawnRadiusOffset + 9));
     }
 }
